Analyzer/requirements.txt
Flask==1.1.1
gunicorn==20.0.4
numpy==1.17.4
dash
pandas_profiling
dash_bootstrap_components
xlrd
flask_caching